Add method to fetch current client state (#26)
## Description

This adds the method `state()` to get the current client state,
including connection status.

If `connection_status.is_good()` returns `true` the client is connected
or trying to reconnect. Otherwise the connection has permanently failed
and the client should be discarded and a new client should be created to
try again later.

We may want to add an asynchronous interface later to get events
whenever the state/status changes.