Improve type safety by avoiding casts (#44)
## Description

This enables hints that warn against the use of casts using the `as`
keyword. Being a low-level library, we need the ability to convert
pointers and references into each other. But instead of doing so with
raw casts (which can do many different things at once), we want to use
explicit methods that describe the specific type of cast we want to
achieve.

This introduces the recently added helpers from the
[`ptr::from_ref()`](https://doc.rust-lang.org/stable/std/ptr/fn.from_ref.html)
family. Combined with pointer methods
[`as_ref()`](https://doc.rust-lang.org/std/primitive.pointer.html#method.as_ref)
and
[`cast()`](https://doc.rust-lang.org/std/primitive.pointer.html#method.cast)
that we already use, we can avoid all raw casts and make them explicit
casts instead.

Unfortunately this requires us to bump the MSRV to 1.76. We might reject
increased type safety for improved compatibility. However, seeing that
there are not many users of this library (yet), I think the increased
ergonomics are worth it.