Simplify return type of call_method() (#79)
## Description

This removes the unnecessary `Option` from the return type of
`AsyncClient::call_method()`. It was used to indicate that the method
returned an unset (invalid) array of output arguments but that should
not happen within the specification.[^1]

[^1]: In any case, it would be unclear what the meaning behind this was
when the call itself was successful.

We treat this as an empty list of output arguments, similar to what we
did for `AsyncClient::browse()` in #77.