Use cycle time as run timeout (#90)
## Description

This is a follow-up to
https://github.com/HMIProject/open62541/pull/84#discussion_r1541989778
and uses the cycle time directly as timeout for
`UA_Client_run_iterate()`.

Previously, we would call that method repeatedly, waiting for cycle time
in-between calls, and use a timeout of zero to avoid blocking. With the
built-in mutex in open62541 1.4, this is no longer necessary: we can set
the desired timeout directly (this is the maximum time to block) and
still issue concurrent requests in the meantime.

The only complication arises during the explicit and implicit shutdown
with `disconnect()` or when the client is dropped. In this case, we need
to make sure to wait for the event loop to finish on its own. We do so
with a cancellation token (atomic bool is sufficient in this case) and
running the background task in its own OS thread.