Remove cycle time from public interface (#91)
## Description

As hinted at in
https://github.com/HMIProject/open62541/pull/90#issuecomment-2039502774,
this removes the `cycle_time` argument from the public interface of
`AsyncClient`.

It is no longer necessary and can be regarded as an implementation
detail: where before this was effectively the polling interval of
handling queued server responses, i.e. the interval we call
`UA_Client_run_iterate()` at, the only user-facing aspect now is the
maximum delay we can expect when dropping a client without calling
`disconnect()` first (which should be avoided anyway).

In other words, the cycle time or iteration timeout only controls the
frequency of cancellation points in our background task: each invocation
of `UA_Client_run_iterate()` blocks and we only must make sure to wake
up frequently enough to be able to interrupt this background loop when
being dropped without `disconnect()`.

Since we can provide a sane default value, this lowers the threshold
when developers first start to work with `AsyncClient`. If we want to
allow fine-tuned values for different environments in the future, we
could make this parameter configurable through `ClientBuilder`.