Add `ServerBuilder` to configure server startup, use custom logger (#95)
## Description

This adds `ServerBuilder` to configure server startup, similar to
`ClientBuilder`. This PR also brings `Client` and `Server` closer
together by mirroring the different ways to construct instances (with
builder or with implied defaults).

It also adds the same logger as used by `Client` to the default `Server`
configuration, to forward log messages through the `log` crate instead
of directly writing to stdout and stderr.