Add documentation to some methods (#118)
## Description

This adds missing documentation to some (internal) methods. In the code,
we use `debug_assert!()` in situations where we expect something that we
control ourselves but nothing too bad would happen if that were not the
case.[^1]

[^1]: This is different from `assert!()` where the situation may happen
spuriously at runtime and we need to include explicit `# Panics`
sections in the documentation.