Add custom payload as context to server nodes (#117)
## Description

This allows upcoming code to associate custom data with server nodes,
including clean-up when the nodes are eventually destroyed by the server
(e.g. when nodes are deleted via #115).

We will use this feature in an upcoming PR to implement variable nodes
with _data source_ support. These data sources will have to hold on to
callbacks as long as the node is alive.

---------

Co-authored-by: Uwe Klotz <uwe.klotz@gmail.com>