Allow defining variable nodes with (callback-driven) `DataSource` (#119)
## Description

This PR introduces callback-driven variable nodes with the use of
`DataSource`. (This is not yet feature-complete, and more work will be
done in upcoming PRs.)

The general idea of data source variable nodes is to allow users to
handle storage manually, in contrast to the existing plain
`Server::add_variable_node()` which handles storage inside.