Avoid destruction of unknown node context for now (#137)
## Description

This PR is a workaround for #125 to prevent segfaults. See that issue
for more details and further investigations to properly solve the
underlying problem.

The changes leak memory: associated data for created nodes is not freed.
As long as the server's node tree is static, this is not a problem. But
we must be aware that repeated node creation and destruction during the
server's runtime will fill up the available heap memory.