Add method nodes (#149)
@sgoll This is a first half-working implementation for adding and using
method nodes. Feel free to leave feedback, if an approach like this is
any good or how it should be changed. Taking a look relatively soon
(let's say until tomorrow lunch time or afternoon) would be appreciated
as then I know if I'm working in the correct direction.

Use case of these methods nodes would be a client calling a callback on
the server that uses some function or does something and then returns
something based on that.

Right now adding the method node works, it can be called in UaExpert. As
far as I could see in the debugger providing parameters didn't work. If
you see any (obvious) errors feel free to leave a comment on that.

Ignore random comments or bad error/panic handling, I will take care of
that tomorrow.

---------

Co-authored-by: Leander Wollersberger <leander.wollersberger@frauscher.com>
Co-authored-by: behrang yousefasr <behrang.yousefasrlangeroodi@frauscher.com>
Co-authored-by: Sebastian Goll <goll@hmi-project.com>