Handle event-notifier monitored items (#222)
## Description

This PR fixes the handling of monitored items with attribute ID
`ua::AttributeId::EVENTNOTIFIER`. For these, the C library expects a
callback function with a different signature (which causes a runtime
crash as described in #218).

This explores an alternative and hopefully simpler implementation that
uses an enum `MonitoredItemValue` to return either the `ua::DataValue`
from regular monitored item notifications or `ua::Array<ua::Variant>`
for event-notifier notifications.

The main disadvantages are

- a low-level cast of function pointers when calling
`UA_Client_MonitoredItems_createDataChanges_async()`,
- more verbose consumption of notification values through the wrapper
enum type `MonitoredItemValue`.

I am not happy with the function pointer magic. Maybe we could mitigate
this through an upstream PR that exports a more general interface that
handles both data change and event notifications (since the attribute ID
is what toggles either of the two callback function signatures, this
would not make the API any more sensitive to misuse as it is).

Regarding the second issue, we could add typed helper methods that
provide basically the same type safety as was shown in #218. We could
make it impossible to (mis-)use these even when we were to allow
`MonitoredItemBuilder` to mix and match regular and event-notifier
monitored items in the same subscription (which is supported by the C
library).