Refactor use of typed `DataValue` (#239)
## Description

This cleans up the invalid assumption that
[`DataValue`](https://reference.opcfoundation.org/Core/Part4/v104/docs/7.7)
only holds a value when `statusCode` has severity Good. This is not
necessarily true, in particular, `DataValue` is allowed to have a value
when `statusCode` is Uncertain, or even Bad.

The implication for our conversion from `ua::DataValue` to
`DataValue<T>` is that we can always do the conversion without errors,
but we must be prepared to deal with unset or unexpected `value`
attributes later.